refactor(query): unify raw SPARQL scanning
